refactor(input): extract spatial grid module
